Fix possible loss of entries depending on sorting.
Using the value to sort as key for an array works only when the key is unique.
While filenames in a folder are unique timestamp and especially file size
aren't, causing files sharing the same information for those getting lost when
sorting after it. Use usort() and a custom comparison function instead to avoid
this.